backend/common/pyproject.toml
[tool.poetry]
name = "common"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [
    { include = "common" }
]

[tool.poetry.dependencies]
python = "^3.11"
celery = {extras = ["redis"], version = "^5.4.0"}
celery-types = "^0.22.0"
elasticsearch_dsl = "^8.13.1"
pymongo = "^4.7.2"
luqum = "^0.13.0"
pycryptodome = "^3.20.0"
requests = "^2.31.0"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
filelock = "^3.14.0"
libretranslatepy = "^2.1.4"
pydantic-mongo = "^2.3.0"
types-requests = "^2.31.0.20240406"
fastapi = "^0.111.0"
types-redis = "^4.6.0.20240425"
ollama = "^0.4.3"
minio = "^7.2.12"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"
freezegun = "^1.5.1"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
